Ensures a successful build on all the 3 frameworks.
