+++ MapboxCarPlay/CarPlayManager.swift
@@ -1,6 +1,7 @@
 #if canImport(CarPlay)
 import CarPlay
 import Turf
+import MapboxNavigation
 import MapboxCoreNavigation
 import MapboxDirections
 
@@ -433,11 +434,8 @@ extension CarPlayManager: CPMapTemplateDelegate {
         let navigationMapTemplate = self.mapTemplate(forNavigating: trip)
         interfaceController.setRootTemplate(navigationMapTemplate, animated: true)
 
-        let navigationViewController = CarPlayNavigationViewController(navigationService: service,
-                                                                       mapTemplate: navigationMapTemplate,
-                                                                       interfaceController: interfaceController,
-                                                                       manager: self,
-                                                                       styles: styles)
+        let navigationViewController = CarPlayNavigationViewController.init(navigationService: service, mapTemplate: navigationMapTemplate, interfaceController: interfaceController, manager: self, styles: styles)
+        
         navigationViewController.startNavigationSession(for: trip)
         navigationViewController.carPlayNavigationDelegate = self
         currentNavigator = navigationViewController
@@ -594,7 +592,7 @@ extension CarPlayManager: CPMapTemplateDelegate {
 
     func coordinate(of offset: CGPoint, in mapView: NavigationMapView) -> CLLocationCoordinate2D {
         
-        let contentFrame = UIEdgeInsetsInsetRect(mapView.bounds, mapView.safeArea)
+        let contentFrame = mapView.bounds.inset(by: mapView.safeArea)
         let centerPoint = CGPoint(x: contentFrame.midX, y: contentFrame.midY)
         let endCameraPoint = CGPoint(x: centerPoint.x - offset.x, y: centerPoint.y - offset.y)
 
@@ -608,7 +606,7 @@ extension CarPlayManager: CPMapTemplateDelegate {
 
         // Determine the screen distance to pan by based on the distance from the visual center to the closest side.
         let mapView = carPlayMapViewController.mapView
-        let contentFrame = UIEdgeInsetsInsetRect(mapView.bounds, mapView.safeArea)
+        let contentFrame = mapView.bounds.inset(by: mapView.safeArea)
         let increment = min(mapView.bounds.width, mapView.bounds.height) / 2.0
         
         // Calculate the distance in physical units from the visual center to where it would be after panning downwards.
