+++ MapboxCarPlay/CarPlayNavigationViewController.swift
@@ -1,7 +1,7 @@
 import Foundation
 import MapboxDirections
 import MapboxCoreNavigation
-#if canImport(CarPlay)
+import MapboxNavigation
 import CarPlay
 
 /**
@@ -456,7 +456,7 @@ public class CarPlayNavigationViewController: UIViewController {
 extension CarPlayNavigationViewController: NavigationMapViewDelegate {
     public func navigationMapViewUserAnchorPoint(_ mapView: NavigationMapView) -> CGPoint {
         // Inset by the content inset to avoid application-defined content.
-        var contentFrame = UIEdgeInsetsInsetRect(mapView.bounds, mapView.contentInset)
+        var contentFrame = mapView.bounds.inset(by: mapView.contentInset)
         
         // Avoid letting the puck go partially off-screen, and add a comfortable padding beyond that.
         let courseViewBounds = mapView.userCourseView?.bounds ?? .zero
@@ -528,4 +528,3 @@ public protocol CarPlayNavigationDelegate {
      */
     @objc func carPlayNavigationViewControllerDidArrive(_ carPlayNavigationViewController: CarPlayNavigationViewController)
 }
-#endif
