+++ MapboxCarPlay/Extensions/UIViewController+CarPlay.swift
@@ -1,6 +1,5 @@
 import UIKit
 
-
 extension UIViewController {
 
     func topMostViewController() -> UIViewController? {
@@ -24,13 +23,13 @@ extension UIViewController {
 }
 
 extension UIWindow {
-    func viewControllerInStack<T: UIViewController>(of type: T.Type? = nil) -> T? {
+     func viewControllerInStack<T: UIViewController>(of type: T.Type? = nil) -> T? {
 
         if let vc = rootViewController as? T {
             return vc
         } else if let vc = rootViewController?.presentedViewController as? T {
             return vc
-        } else if let vc = rootViewController?.childViewControllers {
+        } else if let vc = rootViewController?.children {
             return vc.lazy.compactMap { $0 as? T }.first
         }
