+++ MapboxNavigation/Bundle.swift
@@ -2,7 +2,7 @@ import Foundation
 
 extension Bundle {
     
-    class var mapboxNavigation: Bundle {
+    public class var mapboxNavigation: Bundle {
         get { return Bundle(for: NavigationViewController.self) }
     }
